h5

+++ src/main/java/com/fnjz/front/utils/CreateTokenUtils.java
@@ -4,10 +4,7 @@ import com.alibaba.fastjson.JSON;
 import com.fnjz.commonbean.ResultBean;
 import com.fnjz.constants.ApiResultType;
 import com.fnjz.constants.RedisPrefix;
-import com.fnjz.front.dao.AccountBookRestDao;
-import com.fnjz.front.dao.FengFengTicketRestDao;
-import com.fnjz.front.dao.UserAccountBookRestDao;
-import com.fnjz.front.dao.UserIntegralRestDao;
+import com.fnjz.front.dao.*;
 import com.fnjz.front.entity.api.fengfengticket.FengFengTicketRestEntity;
 import com.fnjz.front.entity.api.userlogin.UserLoginRestEntity;
 import com.fnjz.front.enums.AcquisitionModeEnum;
@@ -43,6 +40,12 @@ public class CreateTokenUtils {
 
     @Autowired
     private UserAccountBookRestDao userAccountBookRestDao;
+
+    @Autowired
+    private UserInviteRestDao userInviteRestDao;
+
+    @Autowired
+    private WarterOrderRestDao warterOrderRestDao;
 
     public String createToken(String code) {
         //使用sharcode作为源token
@@ -103,11 +106,11 @@ public class CreateTokenUtils {
      * @param userInfoId
      * @return
      */
-    public boolean checkTaskComplete(CategoryOfBehaviorEnum categoryOfBehaviorEnum, AcquisitionModeEnum acquisitionModeEnum, String userInfoId,String shareCode) {
+    public boolean checkTaskComplete(CategoryOfBehaviorEnum categoryOfBehaviorEnum, AcquisitionModeEnum acquisitionModeEnum, String userInfoId, String shareCode) {
         //邀请好友可以多次触发
-        if(acquisitionModeEnum.equals(AcquisitionModeEnum.Inviting_friends)){
+        if (acquisitionModeEnum.equals(AcquisitionModeEnum.Inviting_friends)) {
             //缓存邀请好友人数
-            redisTemplateUtils.incrementForHash(RedisPrefix.USER_INVITE_COUNT + shareCode,"inviteCount",1);
+            redisTemplateUtils.incrementForHash(RedisPrefix.USER_INVITE_COUNT + shareCode, "inviteCount", 1);
             return false;
         }
         int count = userIntegralRestDao.checkTaskComplete(categoryOfBehaviorEnum.getIndex(), acquisitionModeEnum.getIndex(), userInfoId);
@@ -120,11 +123,12 @@ public class CreateTokenUtils {
 
     /**
      * 获取行为方式对应的积分数
+     *
      * @param acquisitionModeEnum
      * @return
      */
     public FengFengTicketRestEntity getFengFengTicket(AcquisitionModeEnum acquisitionModeEnum) {
-        return fengFengTicketRestDao.getFengFengTicket(null,acquisitionModeEnum.getName(), null);
+        return fengFengTicketRestDao.getFengFengTicket(null, acquisitionModeEnum.getName(), null);
     }
 
     /**
@@ -134,29 +138,61 @@ public class CreateTokenUtils {
      * @param ff
      * @param acquisitionModeEnum
      */
-    public void insertInIntegral(String userInfoId, FengFengTicketRestEntity ff, AcquisitionModeEnum acquisitionModeEnum,CategoryOfBehaviorEnum categoryOfBehaviorEnum) {
-        userIntegralRestDao.insertSignInIntegral(userInfoId, ff.getId()+"", ff.getBehaviorTicketValue(), acquisitionModeEnum.getDescription(), acquisitionModeEnum.getIndex(),categoryOfBehaviorEnum.getIndex());
-        userIntegralRestDao.updateForTotalIntegral(userInfoId,ff.getBehaviorTicketValue());
+    public void insertInIntegral(String userInfoId, FengFengTicketRestEntity ff, AcquisitionModeEnum acquisitionModeEnum, CategoryOfBehaviorEnum categoryOfBehaviorEnum) {
+        userIntegralRestDao.insertSignInIntegral(userInfoId, ff.getId() + "", ff.getBehaviorTicketValue(), acquisitionModeEnum.getDescription(), acquisitionModeEnum.getIndex(), categoryOfBehaviorEnum.getIndex());
+        userIntegralRestDao.updateForTotalIntegral(userInfoId, ff.getBehaviorTicketValue());
 
     }
 
     /**
      * 引入新手任务
+     *
      * @param categoryOfBehaviorEnum
      * @param acquisitionModeEnum
      */
-    public void integralTask(String userInfoId,String shareCode,CategoryOfBehaviorEnum categoryOfBehaviorEnum,AcquisitionModeEnum acquisitionModeEnum) {
-        boolean flag = this.checkTaskComplete(categoryOfBehaviorEnum, acquisitionModeEnum, userInfoId,shareCode);
+    public void integralTask(String userInfoId, String shareCode, CategoryOfBehaviorEnum categoryOfBehaviorEnum, AcquisitionModeEnum acquisitionModeEnum) {
+        boolean flag = this.checkTaskComplete(categoryOfBehaviorEnum, acquisitionModeEnum, userInfoId, shareCode);
         if (!flag) {
             //获取需要绑定的积分数
             FengFengTicketRestEntity fengFengTicketRestEntity = this.getFengFengTicket(acquisitionModeEnum);
             if (fengFengTicketRestEntity != null) {
-                this.insertInIntegral(userInfoId, fengFengTicketRestEntity, acquisitionModeEnum,categoryOfBehaviorEnum);
-                //删除缓存
-                if(categoryOfBehaviorEnum.equals(CategoryOfBehaviorEnum.NewbieTask)){
-                    redisTemplateUtils.deleteKey(RedisPrefix.PREFIX_NEWBIE_TASK + shareCode);
-                }else if(categoryOfBehaviorEnum.equals(CategoryOfBehaviorEnum.TodayTask)){
-                    redisTemplateUtils.deleteKey(RedisPrefix.PREFIX_TODAY_TASK + shareCode);
+                if (acquisitionModeEnum.equals(AcquisitionModeEnum.The_invitation_came_to_five)) {
+                    //当日邀请达5人  判断人数
+                    int countForInvitedUsers = userInviteRestDao.getCountForInvitedUsers(userInfoId);
+                    if (countForInvitedUsers >= 5) {
+                        this.insertInIntegral(userInfoId, fengFengTicketRestEntity, acquisitionModeEnum, categoryOfBehaviorEnum);
+                        //删除缓存
+                        if (categoryOfBehaviorEnum.equals(CategoryOfBehaviorEnum.NewbieTask)) {
+                            redisTemplateUtils.deleteKey(RedisPrefix.PREFIX_NEWBIE_TASK + shareCode);
+                        } else if (categoryOfBehaviorEnum.equals(CategoryOfBehaviorEnum.TodayTask)) {
+                            redisTemplateUtils.deleteKey(RedisPrefix.PREFIX_TODAY_TASK + shareCode);
+                        }
+                    }
+                } else if (acquisitionModeEnum.equals(AcquisitionModeEnum.The_bookkeeping_came_to_three)) {
+                    //当日记账达3笔
+                    int count = warterOrderRestDao.getCountForCurrentDay(userInfoId);
+                    if (count >= 3) {
+                        this.insertInIntegral(userInfoId, fengFengTicketRestEntity, acquisitionModeEnum, categoryOfBehaviorEnum);
+                        //删除缓存
+                        if (categoryOfBehaviorEnum.equals(CategoryOfBehaviorEnum.NewbieTask)) {
+                            redisTemplateUtils.deleteKey(RedisPrefix.PREFIX_NEWBIE_TASK + shareCode);
+                        } else if (categoryOfBehaviorEnum.equals(CategoryOfBehaviorEnum.TodayTask)) {
+                            redisTemplateUtils.deleteKey(RedisPrefix.PREFIX_TODAY_TASK + shareCode);
+                        }
+                    }
+                } else if (acquisitionModeEnum.equals(AcquisitionModeEnum.Become_hbird_user)) {
+                    //成为蜂鸟记账用户
+                    this.insertInIntegral(userInfoId, fengFengTicketRestEntity, acquisitionModeEnum, categoryOfBehaviorEnum);
+                    //设置注册积分奖励弹框
+                    redisTemplateUtils.cacheForString(RedisPrefix.USER_REGISTER_INTEGRAL + shareCode, fengFengTicketRestEntity.getBehaviorTicketValue()+"",RedisPrefix.USER_VALID_TIME);
+                }else {
+                    this.insertInIntegral(userInfoId, fengFengTicketRestEntity, acquisitionModeEnum, categoryOfBehaviorEnum);
+                    //删除缓存
+                    if (categoryOfBehaviorEnum.equals(CategoryOfBehaviorEnum.NewbieTask)) {
+                        redisTemplateUtils.deleteKey(RedisPrefix.PREFIX_NEWBIE_TASK + shareCode);
+                    } else if (categoryOfBehaviorEnum.equals(CategoryOfBehaviorEnum.TodayTask)) {
+                        redisTemplateUtils.deleteKey(RedisPrefix.PREFIX_TODAY_TASK + shareCode);
+                    }
                 }
             }
         }
@@ -164,18 +200,20 @@ public class CreateTokenUtils {
 
     /**
      * 更新账本最后同步时间
+     *
      * @param abId
      */
-    public void updateABtime(Integer abId){
+    public void updateABtime(Integer abId) {
         accountBookRestDao.updateABtime(abId);
     }
 
     /**
      * 判断当前用户是否具有对传入abId权限
+     *
      * @param abId
      */
-    public boolean checkByABIdAndUserInfoId(Integer abId,String userInfoId){
-        Integer status = userAccountBookRestDao.checkByABIdAndUserInfoId(abId,userInfoId);
-        return status==null?false:true;
+    public boolean checkByABIdAndUserInfoId(Integer abId, String userInfoId) {
+        Integer status = userAccountBookRestDao.checkByABIdAndUserInfoId(abId, userInfoId);
+        return status == null ? false : true;
     }
 }
